Fix FULL_TABLE sync mode to comply with spec
The Singer Spec clearly states that FULL_TABLE replication mode
must replicate "all available records dating back to a `start_date`".

This tap erroneously switches to an incremental replication
strategy if it detects an auto-incrementing primary key. This is
both unsafe (it will omit updates to existing rows) and in violation
of the spec.

Remove this auto-incremental-update logic from the full-table
mode.