Log error message for session SQLs
When setting session SQL such as connection timeouts, it's possible
to get an error like

```
 (1229, "Variable 'connect_timeout' is a GLOBAL variable and should be
 set. with SET GLOBAL")
 ```

The previous error handling just swallowed MySQL error messages,
making it difficult or impossible to debug errors with connection
settings.

Add a simple change to log the DB error message, so that errors
can be diagnosed more easily.