Calculate max/min for DECIMAL columns
Targets must know the width of columns in order to be able to
correctly represent them. For example, BigQuery distinguishes
DECIMAL and BIGDECIMAL types, and a SQL database like target-mysql
or target-postgres would need to be able to reconstruct the actual
precision/scale values in order to be able to create a valid schema.

To encode the DECIMAL column as JSONSchema, use Python Decimal
machinery to calculate the Maximum/Minimum values of a decimal
with the specified precision.