Allow schema overrides
Passing a catalog file with some schema overrides, e.g. adding `format`
to a string field to inform the target about the specifig type of
string, wasn't supported.

The reason was that the sync process re-discovers the catalog even
when an input file is provided in order to prevent selection of items
not present in the database.